+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Times dropdown must be above header in fixed position 
 * Admin: Better label on change booking status button
+* Front: Only load times when on step 2
 
 ### 1.4.2: May 19th, 2021
 * No Distraction for StormBringer theme