+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Admin: Better label on change booking status button
 * Front: Only load times when on step 2
 * Fix no distraction mode after Elementor footer update
+* Enqueue min JS file for production
 
 ### 1.4.2: May 19th, 2021
 * No Distraction for StormBringer theme