+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix no distraction mode after Elementor footer update
 * Enqueue min JS file for production
 * New form field in add form: summary to display selected summary of the booking before confirming
+* Smooth transition to summary field after selecting time
 
 ### 1.4.2: May 19th, 2021
 * No Distraction for StormBringer theme