+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.4.3: May 24th, 2021
 * Times dropdown must be above header in fixed position 
 * Admin: Better label on change booking status button
 * Front: Only load times when on step 2
@@ -5,7 +6,6 @@
 * Enqueue min JS file for production
 * New form field in add form: summary to display selected summary of the booking before confirming
 * Smooth transition to summary field after selecting time
-* Let Gravity Forms hide/show the confirm button
 
 ### 1.4.2: May 19th, 2021
 * No Distraction for StormBringer theme